feat: 整理 where cum_sum shape slice
