+++ src/03onnx/src/infer/infer.cpp
@@ -312,157 +312,6 @@ namespace refactor::onnx {
         }
     }
 
-    InferResult inferCumSum(Operator const &op, Edges inputs) {
-        EXPECT_SIZE(2)
-        if (!inputs[1]->shape.empty()) {
-            return Err(InferError(ERROR_MSG("Input shape not support")));
-        } else if (!isNumbericDataType(inputs[0]->dataType) ||
-                   (inputs[1]->dataType != DataType::I64 &&
-                    inputs[1]->dataType != DataType::I32)) {
-            return Err(InferError(ERROR_MSG("Input data type not support")));
-        } else {
-            return Ok(Edges{std::move(inputs[0])});
-        }
-    }
-
-    InferResult inferSlice(Operator const &op, Edges inputs) {
-        if (inputs.size() < 3 || 5 < inputs.size()) {
-            return Err(InferError(ERROR_MSG("Input size error")));
-        } else {
-            auto const &data = inputs[0];
-            auto const &starts_ = inputs[1];
-            auto const &ends_ = inputs[2];
-            auto tint = starts_->dataType;
-            if (tint != DataType::I32 && tint != DataType::I64) {
-                return Err(InferError(ERROR_MSG("Data type not support")));
-            }
-            if (ends_->dataType != tint) {
-                return Err(InferError(ERROR_MSG("Data type not support")));
-            }
-            if (starts_->shape.size() != 1) {
-                return Err(InferError(ERROR_MSG("Input shape not support")));
-            }
-            if (ends_->shape != starts_->shape) {
-                return Err(InferError(ERROR_MSG("Input shape not support")));
-            }
-            EXPECT_VAL(starts_->shape[0], rank)
-            if (data->shape.size() != rank) {
-                return Err(InferError(ERROR_MSG("Input shape not support")));
-            }
-            if (!starts_->hasData() || !ends_->hasData()) {
-                return Err(InferError(ERROR_MSG("Starts and ends must be constant")));
-            }
-            int64_t *starts = reinterpret_cast<int64_t *>(starts_->data->ptr),
-                    *ends = reinterpret_cast<int64_t *>(ends_->data->ptr),
-                    *axes = nullptr,
-                    *steps = nullptr;
-            std::vector<int64_t> axes__, steps__;
-            if (inputs.size() < 4) {
-                axes__.resize(rank);
-                axes = axes__.data();
-                for (int64_t i = 0; i < rank; ++i) {
-                    axes[i] = i;
-                }
-            } else {
-                auto const &axes_ = inputs[3];
-                if (axes_->dataType != tint || axes_->shape != starts_->shape) {
-                    return Err(InferError(ERROR_MSG("Axes not support")));
-                }
-                if (!axes_->hasData()) {
-                    return Err(InferError(ERROR_MSG("Axes must be constant")));
-                }
-                axes = reinterpret_cast<int64_t *>(axes_->data->ptr);
-            }
-            if (inputs.size() < 5) {
-                steps__.resize(rank, 1);
-                steps = steps__.data();
-            } else {
-                auto const &steps_ = inputs[4];
-                if (steps_->dataType != tint || steps_->shape != starts_->shape) {
-                    return Err(InferError(ERROR_MSG("Steps not support")));
-                }
-                if (!steps_->hasData()) {
-                    return Err(InferError(ERROR_MSG("Steps must be constant")));
-                }
-                steps = reinterpret_cast<int64_t *>(steps_->data->ptr);
-            }
-
-            Shape ans(rank, DimExpr(1));
-            for (size_t i = 0; i < rank; ++i) {
-                auto axis = axes[i];
-                auto start = starts[i];
-                auto end = ends[i];
-                auto step = steps[i];
-                if (axis < 0) {
-                    axis += rank;
-                }
-                EXPECT_VAL(data->shape[axis], dim)
-                if (start < 0) {
-                    start += dim;
-                }
-                if (end < 0) {
-                    end += dim;
-                }
-                if (start < 0 || dim <= start || end < 0 || dim < end) {
-                    return Err(InferError(ERROR_MSG("Input shape not support")));
-                }
-                if (step > 0) {
-                    ans[axis] = DimExpr((end - start + step - 1) / step);
-                } else if (step < 0) {
-                    ans[axis] = DimExpr((end - start - step + 1) / -step);
-                } else {
-                    return Err(InferError(ERROR_MSG("Input shape not support")));
-                }
-            }
-            return Ok(Edges{std::make_shared<Tensor>(data->dataType, std::move(ans))});
-        }
-    }
-
-    InferResult inferShape(Operator const &op, Edges inputs) {
-        EXPECT_SIZE(1) {
-            auto attrs = op.attributes;
-            auto start = op.attribute("start", {0}).int_(),
-                 end = op.attribute("end", {-1}).int_();
-
-            ASSERT(start == 0, "only support start == 0");
-            ASSERT(end == -1, "only support end == -1");
-
-            auto ans = inputs[0]->shape;
-            auto blob = std::make_shared<Blob>(new int64_t[ans.size()]);
-            for (auto i = 0; i < ans.size(); ++i) {
-                EXPECT_VAL(ans[i], v)
-                reinterpret_cast<int64_t *>(blob->ptr)[i] = v;
-            }
-            return Ok(Edges{
-                std::make_shared<Tensor>(
-                    DataType::I64,
-                    Shape{DimExpr(ans.size())},
-                    std::move(blob)),
-            });
-        }
-    }
-
-    InferResult inferWhere(Operator const &op, Edges inputs) {
-        EXPECT_SIZE(3) {
-            auto const &condition = inputs[0];
-            auto const &x = inputs[1];
-            auto const &y = inputs[2];
-            if (condition->dataType != DataType::Bool) {
-                return Err(InferError(ERROR_MSG("Input data type not support")));
-            }
-            if (x->dataType != y->dataType) {
-                return Err(InferError(ERROR_MSG("Input data type not support")));
-            }
-            auto ans = multidirBroadcast({condition->shape, x->shape, y->shape});
-            if (ans.isErr()) {
-                return Err(InferError(ERROR_MSG(ans.unwrapErr())));
-            } else {
-                return Ok(Edges{std::make_shared<Tensor>(x->dataType, ans.unwrap())});
-            }
-        }
-    }
-
-
     InferResult inferSoftmax(Operator const &op, Edges inputs) {
         EXPECT_SIZE(1)
         if (!isIeee754DataType(inputs[0]->dataType)) {
