+++ src/03onnx/src/infer/where.cpp
@@ -1,0 +1,25 @@
+﻿#include "infer.h"
+
+namespace refactor::onnx {
+    using namespace refactor::common;
+
+    InferResult inferWhere(Operator const &op, Edges inputs) {
+        EXPECT_SIZE(3) {
+            auto const &condition = inputs[0];
+            auto const &x = inputs[1];
+            auto const &y = inputs[2];
+            if (condition->dataType != DataType::Bool) {
+                return Err(InferError(ERROR_MSG("Input data type not support")));
+            }
+            if (x->dataType != y->dataType) {
+                return Err(InferError(ERROR_MSG("Input data type not support")));
+            }
+            auto ans = multidirBroadcast({condition->shape, x->shape, y->shape});
+            if (ans.isErr()) {
+                return Err(InferError(ERROR_MSG(ans.unwrapErr())));
+            } else {
+                return Ok(Edges{std::make_shared<Tensor>(x->dataType, ans.unwrap())});
+            }
+        }
+    }
+}// namespace refactor::onnx
